issue #124 refactor: merge calendar constants in src/constants/calendarConstants and fix related import statement
