+++ src/pages/PersonalSchedulePage/PersonalSchedulePage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import CalendarContainer from "@/components/Common/CalendarContainer";
-import SCHEDULE_TYPE from "@/constants/calendar/scheduleType";
+import { SCHEDULE_TYPE } from "@/constants/calendarConstants";
 
 import { LayoutMain } from "./PersonalSchedulePage.styles";
 import PersonalTodoList from "../../components/PersonalTodoList/PersonalTodoList/PersonalTodoList";
@@ -9,7 +9,7 @@ import PersonalTodoList from "../../components/PersonalTodoList/PersonalTodoList
 const PersonalSchedulePage = () => {
 	return (
 		<LayoutMain>
-			<CalendarContainer type={SCHEDULE_TYPE.personal} />
+			<CalendarContainer type={SCHEDULE_TYPE.PERSONAL} />
 			<PersonalTodoList />
 		</LayoutMain>
 	);
